Use checkbox menu in crate selection.
Switch from normal QAction menu to a composition of QCheckBox and QActionWidget.
The checkbox shows in which crates the selection is in.
Changing the crates selection does not collapse the menu, which allows
much easier categorization of tracks without going through the menu from scratch.

Use tristate when multiple tracks are selected.

When mulitple tracks are selected which do not share a crate, use
the tristate partially selected to indicate this.

Fix styling of pointers.

Use QVariant to transport CrateId in checkbox